+++ flutter/lib/src/native/ffi/exception.g.dart
@@ -37,13 +37,27 @@ void registerFunction(
       newMediaSettingsUpdateException,
   required Pointer<NativeFunction<Handle Function(Pointer<Utf8>)>>
       newInvalidOutputAudioDeviceIdException,
+  required Pointer<
+          NativeFunction<Handle Function(Pointer<Handle>, Pointer<Utf8>)>>
+      newMicVolumeException,
   required Pointer<NativeFunction<Handle Function()>> throwPanicException,
 }) {
   dl.lookupFunction<
       Void Function(Pointer, Pointer, Pointer, Pointer, Pointer, Pointer,
-          Pointer, Pointer, Pointer, Pointer, Pointer),
-      void Function(Pointer, Pointer, Pointer, Pointer, Pointer, Pointer,
-          Pointer, Pointer, Pointer, Pointer, Pointer)>('register_exception')(
+          Pointer, Pointer, Pointer, Pointer, Pointer, Pointer),
+      void Function(
+          Pointer,
+          Pointer,
+          Pointer,
+          Pointer,
+          Pointer,
+          Pointer,
+          Pointer,
+          Pointer,
+          Pointer,
+          Pointer,
+          Pointer,
+          Pointer)>('register_exception')(
     newArgumentError,
     newStateError,
     newFormatException,
@@ -54,6 +68,7 @@ void registerFunction(
     newInternalException,
     newMediaSettingsUpdateException,
     newInvalidOutputAudioDeviceIdException,
+    newMicVolumeException,
     throwPanicException,
   );
 }