+++ flutter/lib/src/native/platform/media_devices.g.dart
@@ -9,16 +9,24 @@ void registerFunction(
   required Pointer<NativeFunction<Handle Function(Handle)>> getDisplayMedia,
   required Pointer<NativeFunction<Handle Function(Pointer<Utf8>)>>
       setOutputAudioId,
+  required Pointer<NativeFunction<Handle Function()>>
+      microphoneVolumeIsAvailable,
+  required Pointer<NativeFunction<Handle Function()>> microphoneVolume,
+  required Pointer<NativeFunction<Handle Function(Int64)>> setMicrophoneVolume,
   required Pointer<NativeFunction<Void Function(Handle)>> onDeviceChange,
 }) {
   dl.lookupFunction<
-      Void Function(Pointer, Pointer, Pointer, Pointer, Pointer),
-      void Function(Pointer, Pointer, Pointer, Pointer,
-          Pointer)>('register_media_devices')(
+      Void Function(Pointer, Pointer, Pointer, Pointer, Pointer, Pointer,
+          Pointer, Pointer),
+      void Function(Pointer, Pointer, Pointer, Pointer, Pointer, Pointer,
+          Pointer, Pointer)>('register_media_devices')(
     enumerateDevices,
     getUserMedia,
     getDisplayMedia,
     setOutputAudioId,
+    microphoneVolumeIsAvailable,
+    microphoneVolume,
+    setMicrophoneVolume,
     onDeviceChange,
   );
 }